requirements.txt
python-nmap
requests
msgpack-python
pymetasploit3
openai
httpx
